Moving font types to a separate file
