+++ src/index.ts
@@ -1,6 +1,6 @@
 import Char from './components/Character';
 import Phrase from './components/Phrase';
-import { SvgChar } from './components/fonts/index';
+import { SvgChar } from './types/font';
 
 export { Phrase, Char };
 export type { SvgChar };
